# api/kube-jit/docs/swagger.yaml
basePath: /kube-jit-api
definitions:
...
  models.LoginResponse:
    properties:
      expiresIn:
        type: integer
      userData:
        $ref: '#/definitions/models.NormalizedUserData'
    type: object
...
  models.NormalizedUserData:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
  models.SimpleMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Self-service Kubernetes RBAC JIT Requests with Google/Azure/Github
    Oauth.
  title: Kube-JIT API
  version: "1.0"
paths:
...
  /kube-jit-api/github/profile:
    get:
      consumes:
      - application/json
      description: |-
        Returns the normalized GitHub user profile for the authenticated user.
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NormalizedUserData'
        "401":
          description: 'Unauthorized: no token in session data'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Get the logged in user's GitHub profile
      tags:
      - github
...
  /kube-jit-api/oauth/github/callback:
    get:
      consumes:
      - application/json
      description: Handles the GitHub OAuth callback, exchanges the code for an access
        token, fetches user info, sets session data, and returns normalized user data
        and expiration time.
      parameters:
      - description: GitHub OAuth authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Normalized user data and expiration time
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Missing or invalid code
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "403":
          description: Unauthorized org
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: GitHub OAuth callback
      tags:
      - github
...
swagger: "2.0"